split out the rest of the routes
